requirements.txt
torch
torchvision
numpy
opencv
opencv-contrib-python
scikit-image
pytorch-msssim
scipy
pillow
